Add CMake to build Glimmer Library   (#4)
* CMake Changes for Build setup

* Update BUILD.md

Updated Build.md

* Added Build Dependency script file

* Build Fix

* Update build_dependencies

* Disable certain flags

---------

+++ GlimmerTest/GlimmerTestGLFW/platform.cpp
@@ -215,9 +215,9 @@ namespace glimmer
                 ImGui_ImplOpenGL3_NewFrame();
                 ImGui_ImplGlfw_NewFrame();
 
-                if (EnterFrame(width, height, CustomEventData{}))
+                if (EnterFrame(static_cast<float>(width), static_cast<float>(height), CustomEventData{}))
                 {
-                    close = !runner(ImVec2{ width, height }, *this, data);
+                    close = !runner(ImVec2{ static_cast<float>(width), static_cast<float>(height) }, *this, data);
 
                     for (auto [data, handler] : handlers)
                         close = !handler(data, desc) && close;
